refactor: separate render logic and add more doc comments
