+++ epub/ncx.go
@@ -7,10 +7,13 @@ import (
 
 const idNCX = "ncx"
 
+// NCX represents an EPUB 2.0 compatible navigation document.
 type NCX struct {
 	NavPoints []NavPoint `xml:"navMap>navPoint"`
 }
 
+// NavPoint represents a location within the epub file that can be navigated
+// to.
 type NavPoint struct {
 	ID        string `xml:"id,attr"`
 	PlayOrder string `xml:"playOrder,attr"`
@@ -50,6 +53,7 @@ func (r *Reader) setNCX() error {
 	return nil
 }
 
+// ncxItemName searches for the name of an item in an NCX document.
 func (rf Rootfile) ncxItemName(href string) string {
 	for _, point := range rf.NCX.NavPoints {
 		if label := point.lookupItemName(href); label != "" {
@@ -60,6 +64,7 @@ func (rf Rootfile) ncxItemName(href string) string {
 	return ""
 }
 
+// lookupItemName traverses a NavPoint looking for the name of an item.
 func (np NavPoint) lookupItemName(href string) string {
 	if np.Content.Src == href {
 		return np.NavLabel.Text
